+++ src/directives/gauge.js
@@ -152,6 +152,6 @@
                     animationHelpers[scope.displayMode].setup();
                 });
             }
-        }
+        };
     }]);
 })();