+++ src/functions.js
@@ -603,6 +603,109 @@
                 }
                 return ratio;
             }
+        },
+        DateFieldUtils: {
+            datePatternBuilder: function (mode) {
+                var patterns = {
+                    highcharts: {
+                        'Hh': '%Hh', // '00h', '01h', ... '23h'
+                        'MMM': '%M', // 'Jan', 'Feb', ... 'Dec'
+                        'YYYY': '%Y', // '2011', '2012', '2013'...
+                        'MMMM': '%B', // 'January', 'February', ... 'December'
+                        'D': '%e', // '1', '2', ... '31',
+                        'ddd': '%a' // 'Sun', 'Mon', ... 'Sat'
+                    },
+                    moment: {
+                        'Hh': 'H[h]',
+                        'MMM': 'MMM',
+                        'YYYY': 'YYYY',
+                        'MMMM': 'MMMM',
+                        'D': 'D',
+                        'ddd': 'ddd'
+                    }
+                }[mode];
+                
+                return function (object) {
+                    var datePattern = '';
+                    if (angular.isObject(object) && ('year' in object || 'month' in object || 'day' in object || 'hour' in object || 'minute' in object || 'weekday' in object)) {
+                        if (!('year' in object)) {
+                            if ('month' in object) {
+                                datePattern = patterns['MMMM'];
+                            }
+                            if ('day' in object) {
+                                if ('month' in object) {
+                                    datePattern = patterns['D'] + ' ' + patterns['MMMM'];
+                                } else {
+                                    datePattern = patterns['D'];
+                                }
+                            }
+                            if ('weekday' in object) {
+                                datePattern = patterns['ddd'];
+                                if ('hour' in object) {
+                                    datePattern += ' ' + patterns['Hh'];
+                                }
+                            } else if ('hour' in object) {
+                                datePattern = patterns['Hh'];
+                            }
+                        } else {
+                            if ('day' in object) {
+                                datePattern += ' ' + patterns['D'];
+                            }
+                            if ('month' in object) {
+                                datePattern += ' ' + patterns['MMMM'];
+                            }
+                            datePattern += ' ' + patterns['YYYY'];
+
+                            if ('hour' in object) {
+                                if ('minute' in object) {
+                                    datePattern += ' ' + patterns['Hh'] + patterns['MMM'];
+                                } else {
+                                    datePattern += ' ' + patterns['Hh'];
+                                }
+                            }
+                        }
+                    }
+                    return datePattern;
+                }
+            },
+            getDateFromXObject: function (x, minDate) {
+                var minYear = minDate ? minDate.getUTCFullYear() : 2000;
+                var minMonth = minDate ? minDate.getUTCMonth() : 0;
+                var minDay = minDate ? minDate.getUTCDate() : 1;
+                var minHour = minDate ? minDate.getUTCHours() : 0;
+                var minMinute = minDate ? minDate.getUTCMinutes() : 0;
+
+                if (angular.isObject(x) && ('year' in x || 'month' in x || 'day' in x || 'hour' in x || 'minute' in x || 'weekday' in x || 'yearday' in x)) {
+                    // default to 2000 because it's a leap year
+                    var date = new Date(Date.UTC(x.year || minYear, x.month - 1 || 0, x.day || 1, x.hour || 0, x.minute || 0));
+                    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date#Two digit years
+                    date.setUTCFullYear(x.year || minYear);
+                    if (!('month' in x)) date.setUTCMonth(minMonth);
+                    if (!('day' in x)) date.setUTCDate(minDay);
+                    if (!('hour' in x)) date.setUTCHours(minHour);
+                    if (!('minute' in x)) date.setUTCMinutes(minMinute);
+                    if (!('year' in x)) {
+                        if ('weekday' in x) {
+                            date.setUTCDate(date.getUTCDate() + 7 - date.getUTCDay() + x.weekday);
+                        }
+                        if ('yearday' in x) {
+                            date.setUTCDate(0 + x.yearday);
+                        }
+                    }
+                    if ('day' in x) {
+                        // handle bisextil years
+                        if (x.day == 29 && x.month == 2 && !x.year) {
+                            date.setUTCDate(28);
+                            date.setUTCMonth(1);
+                        }
+                    } else {
+                        if ('month' in x) {
+                            date.setUTCDate(16);
+                        }
+                    }
+                    return date;
+                }
+            }
         }
     };